Update README.md
Enhanced the README to be more descriptive and informative, reflecting a comprehensive guide for beginners in Python face recognition. Added detailed steps for installation, library usage, and practical implementation. Integrated a section on DeepFace for advanced facial analysis. Included a new segment on Dockerizing the project, providing instructions for containerization to simplify deployment and ensure consistent environments. This update aims to make the project more accessible and user-friendly, catering to both beginners and those looking to scale or deploy their applications efficiently.